refactor: flatten structure of config module
- move tests into combined test module
- move keymap submodules into config module
- AppConfig now needs a config directory passed into it
