+++ src/config/tests/mod.rs
@@ -1,46 +1,52 @@
-use std::path::PathBuf;
-
 use super::*;
 
-#[test]
-fn test_parse_key_event() -> Result<()> {
-    let test_cases = [
-        (
-            "ctrl-c",
-            KeyEvent::new(KeyCode::Char('c'), KeyModifiers::CONTROL),
-            "Test lowercase string",
-        ),
-        (
-            "CTRL-C",
-            KeyEvent::new(KeyCode::Char('c'), KeyModifiers::CONTROL),
-            "Test all caps string",
-        ),
-        (
-            "Ctrl-C",
-            KeyEvent::new(KeyCode::Char('c'), KeyModifiers::CONTROL),
-            "Test variable capitalization",
-        ),
-        (
-            "alt-tab",
-            KeyEvent::new(KeyCode::Tab, KeyModifiers::ALT),
-            "Test modifier and non-char keycode",
-        ),
-    ];
+mod parsing;
 
-    test_cases.iter().try_for_each(|(raw, want, desc)| {
-        Aok::<()>({
-            let got = parse_key_event(raw)?;
-            assert_eq!(*want, got, "{desc}");
-        })
-    })?;
+impl AppConfig {
+    fn mock_root() -> Result<PathBuf> {
+        Ok(PathBuf::from_iter([
+            env::var("PWD")?.as_str(),
+            ".config",
+            "mathing",
+        ]))
+    }
+    fn mock_db_file() -> Result<PathBuf> {
+        Ok(Self::mock_root()?.join("data.db"))
+    }
+    fn mock_keymap_file() -> Result<PathBuf> {
+        Ok(Self::mock_root()?.join("config.toml"))
+    }
+}
+
+#[tokio::test]
+async fn test_db_init() -> Result<()> {
+    let conn = DbConn::try_init(AppConfig::mock_db_file()?).await;
+    assert!(conn.is_ok(), "Test db connection initialization.");
 
     Ok(())
 }
 
 #[test]
+fn test_keymap_init() -> Result<()> {
+    let keymap = KeyMap::try_init(AppConfig::mock_keymap_file()?);
+    assert!(keymap.is_ok(), "Test keymap initialization.");
+
+    Ok(())
+}
+
+#[test]
+fn test_get_config_dir() -> Result<()> {
+    // test different env variables
+    assert!(
+        false,
+        "Test getting the config dir with the different env varaibles."
+    );
+    Ok(())
+}
+
+#[test]
 fn test_keymap_builder() -> Result<()> {
-    let config_dir = PathBuf::from_iter([".config", "mathing", "config.toml"]);
-    let keymap = KeyMap::try_init(config_dir)?;
+    let keymap = KeyMap::try_init(AppConfig::mock_keymap_file()?)?;
 
     let test_cases = [
         (
